# vulkan: optimize coopmat2 q2_k dequant function