Same kind of optimization as in https://github.com/ggerganov/llama.cpp/pull/10855, just for Q2_K. I happened to be trying out a Q2_K model for stable-diffusion/flux and this makes it about 5% faster.

```
stable-diffusion:
sd --diffusion-model  models\flux\flux1-dev-Q2_K.gguf --vae models\flux\ae.safetensors --clip_l models\flux\clip_l.safetensors --t5xxl models\flux\t5xxl_fp16.safetensors -p "a lovely cat holding a sign says 'flux.cpp'" --cfg-scale 1.0 --sampling-method euler -v --diffusion-fa
coopmat2 before:
  |==================================================| 20/20 - 1.45it/s
coopmat2 after:
  |==================================================| 20/20 - 1.53it/s
coopmat2 after without --diffusion-fa:
  |==================================================| 20/20 - 1.20it/s
coopmat1 without --diffusion-fa:
  |==================================================| 20/20 - 1.08it/s
no coopmat without --diffusion-fa:
  |==================================================| 20/20 - 1.83s/it ( == 0.55it/s)

test-backend-ops:
before
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 2880 runs -  1736.50 us/run -  60.13 GFLOP/run -  34.63 TFLOPS
after
  MUL_MAT(type_a=q2_K,type_b=f32,m=4096,n=512,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                 3094 runs -  1616.07 us/run -  60.13 GFLOP/run -  37.21 TFLOPS
```